Default implementation of UserControllerType
